requirements.txt
flask
crewai
crewai[google-genai]
crewai-tools
python-dotenv
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
